refactor: extract persist snapshot helpers
