+++ src/persist.rs
@@ -8,270 +8,22 @@ use std::sync::atomic::AtomicBool;
 use std::sync::Arc;
 
 use ratatui::layout::Direction;
-use serde::{Deserialize, Serialize};
-use tracing::{error, info, warn};
-
 use tokio::sync::{mpsc, Notify};
+use tracing::{error, info, warn};
 
 use crate::events::AppEvent;
 use crate::layout::{Node, PaneId, TileLayout};
 use crate::pane::{PaneRuntime, PaneState};
 use crate::workspace::Workspace;
 
-/// Current snapshot format version.
-const SNAPSHOT_VERSION: u32 = 3;
+mod snapshot;
 
-/// Serializable snapshot of the entire herdr session.
-#[derive(Serialize, Deserialize)]
-pub struct SessionSnapshot {
-    /// Format version — used to detect incompatible changes.
-    #[serde(default)]
-    pub version: u32,
-    pub workspaces: Vec<WorkspaceSnapshot>,
-    pub active: Option<usize>,
-    pub selected: usize,
-    #[serde(default)]
-    pub agent_panel_scope: crate::app::state::AgentPanelScope,
-    #[serde(default)]
-    pub sidebar_width: Option<u16>,
-    #[serde(default)]
-    pub sidebar_section_split: Option<f32>,
-}
-
-#[derive(Serialize, Deserialize)]
-pub struct WorkspaceSnapshot {
-    #[serde(default)]
-    pub id: Option<String>,
-    #[serde(default)]
-    pub custom_name: Option<String>,
-    pub identity_cwd: PathBuf,
-    pub tabs: Vec<TabSnapshot>,
-    #[serde(default)]
-    pub active_tab: usize,
-}
-
-#[derive(Deserialize)]
-struct LegacyWorkspaceSnapshot {
-    #[serde(default)]
-    custom_name: Option<String>,
-    layout: LayoutSnapshot,
-    panes: HashMap<u32, PaneSnapshot>,
-    zoomed: bool,
-    #[serde(default)]
-    focused: Option<u32>,
-    #[serde(default)]
-    root_pane: Option<u32>,
-}
-
-#[derive(Serialize, Deserialize)]
-pub struct TabSnapshot {
-    #[serde(default)]
-    pub custom_name: Option<String>,
-    pub layout: LayoutSnapshot,
-    pub panes: HashMap<u32, PaneSnapshot>,
-    pub zoomed: bool,
-    #[serde(default)]
-    pub focused: Option<u32>,
-    #[serde(default)]
-    pub root_pane: Option<u32>,
-}
-
-#[derive(Serialize, Deserialize)]
-pub struct PaneSnapshot {
-    pub cwd: PathBuf,
-}
-
-/// Serializable BSP tree.
-#[derive(Serialize, Deserialize)]
-pub enum LayoutSnapshot {
-    Pane(u32),
-    Split {
-        direction: DirectionSnapshot,
-        ratio: f32,
-        first: Box<LayoutSnapshot>,
-        second: Box<LayoutSnapshot>,
-    },
-}
-
-#[derive(Serialize, Deserialize)]
-pub enum DirectionSnapshot {
-    Horizontal,
-    Vertical,
-}
-
-impl From<LegacyWorkspaceSnapshot> for WorkspaceSnapshot {
-    fn from(snap: LegacyWorkspaceSnapshot) -> Self {
-        let identity_cwd = legacy_identity_cwd(&snap);
-        let tab = TabSnapshot {
-            custom_name: None,
-            layout: snap.layout,
-            panes: snap.panes,
-            zoomed: snap.zoomed,
-            focused: snap.focused,
-            root_pane: snap.root_pane,
-        };
-
-        Self {
-            id: None,
-            custom_name: snap.custom_name,
-            identity_cwd,
-            tabs: vec![tab],
-            active_tab: 0,
-        }
-    }
-}
-
-#[derive(Deserialize)]
-struct RawSessionSnapshot {
-    #[serde(default)]
-    version: u32,
-    #[serde(default)]
-    workspaces: Vec<serde_json::Value>,
-    #[serde(default)]
-    active: Option<usize>,
-    #[serde(default)]
-    selected: usize,
-    #[serde(default)]
-    agent_panel_scope: crate::app::state::AgentPanelScope,
-    #[serde(default)]
-    sidebar_width: Option<u16>,
-    #[serde(default)]
-    sidebar_section_split: Option<f32>,
-}
-
-fn migrate_snapshot(raw: RawSessionSnapshot) -> Result<SessionSnapshot, String> {
-    Ok(SessionSnapshot {
-        version: raw.version,
-        workspaces: raw
-            .workspaces
-            .into_iter()
-            .map(migrate_workspace)
-            .collect::<Result<Vec<_>, _>>()?,
-        active: raw.active,
-        selected: raw.selected,
-        agent_panel_scope: raw.agent_panel_scope,
-        sidebar_width: raw.sidebar_width,
-        sidebar_section_split: raw.sidebar_section_split,
-    })
-}
-
-fn migrate_workspace(raw: serde_json::Value) -> Result<WorkspaceSnapshot, String> {
-    if raw.get("identity_cwd").is_some() {
-        return serde_json::from_value(raw).map_err(|e| e.to_string());
-    }
-
-    if raw.get("layout").is_some() {
-        let legacy =
-            serde_json::from_value::<LegacyWorkspaceSnapshot>(raw).map_err(|e| e.to_string())?;
-        return Ok(legacy.into());
-    }
-
-    Err("workspace snapshot is neither current nor legacy format".to_string())
-}
-
-fn legacy_identity_cwd(snap: &LegacyWorkspaceSnapshot) -> PathBuf {
-    let root_pane = snap
-        .root_pane
-        .or_else(|| first_pane_id_in_layout(&snap.layout));
-
-    root_pane
-        .and_then(|pane_id| snap.panes.get(&pane_id))
-        .map(|pane| pane.cwd.clone())
-        .or_else(|| {
-            first_pane_id_in_layout(&snap.layout)
-                .and_then(|pane_id| snap.panes.get(&pane_id))
-                .map(|pane| pane.cwd.clone())
-        })
-        .or_else(|| {
-            snap.panes
-                .keys()
-                .min()
-                .and_then(|pane_id| snap.panes.get(pane_id))
-                .map(|pane| pane.cwd.clone())
-        })
-        .unwrap_or_else(|| std::env::current_dir().unwrap_or_else(|_| "/".into()))
-}
-
-fn first_pane_id_in_layout(layout: &LayoutSnapshot) -> Option<u32> {
-    match layout {
-        LayoutSnapshot::Pane(id) => Some(*id),
-        LayoutSnapshot::Split { first, second, .. } => {
-            first_pane_id_in_layout(first).or_else(|| first_pane_id_in_layout(second))
-        }
-    }
-}
-
-// --- Capture ---
-
-/// Capture the current app state into a serializable snapshot.
-pub fn capture(
-    workspaces: &[Workspace],
-    active: Option<usize>,
-    selected: usize,
-    agent_panel_scope: crate::app::state::AgentPanelScope,
-    sidebar_width: u16,
-    sidebar_section_split: f32,
-) -> SessionSnapshot {
-    SessionSnapshot {
-        version: SNAPSHOT_VERSION,
-        workspaces: workspaces.iter().map(capture_workspace).collect(),
-        active,
-        selected,
-        agent_panel_scope,
-        sidebar_width: Some(sidebar_width),
-        sidebar_section_split: Some(sidebar_section_split),
-    }
-}
-
-fn capture_workspace(ws: &Workspace) -> WorkspaceSnapshot {
-    WorkspaceSnapshot {
-        id: Some(ws.id.clone()),
-        custom_name: ws.custom_name.clone(),
-        identity_cwd: ws
-            .resolved_identity_cwd()
-            .unwrap_or_else(|| ws.identity_cwd.clone()),
-        tabs: ws.tabs.iter().map(capture_tab).collect(),
-        active_tab: ws.active_tab,
-    }
-}
-
-fn capture_tab(tab: &crate::workspace::Tab) -> TabSnapshot {
-    let mut panes = HashMap::new();
-    for id in tab.panes.keys() {
-        let cwd = tab
-            .cwd_for_pane(*id)
-            .unwrap_or_else(|| std::env::current_dir().unwrap_or_else(|_| "/".into()));
-        panes.insert(id.raw(), PaneSnapshot { cwd });
-    }
-    TabSnapshot {
-        custom_name: tab.custom_name.clone(),
-        layout: capture_node(tab.layout.root()),
-        panes,
-        zoomed: tab.zoomed,
-        focused: Some(tab.layout.focused().raw()),
-        root_pane: Some(tab.root_pane.raw()),
-    }
-}
-
-fn capture_node(node: &Node) -> LayoutSnapshot {
-    match node {
-        Node::Pane(id) => LayoutSnapshot::Pane(id.raw()),
-        Node::Split {
-            direction,
-            ratio,
-            first,
-            second,
-        } => LayoutSnapshot::Split {
-            direction: match direction {
-                Direction::Horizontal => DirectionSnapshot::Horizontal,
-                Direction::Vertical => DirectionSnapshot::Vertical,
-            },
-            ratio: *ratio,
-            first: Box::new(capture_node(first)),
-            second: Box::new(capture_node(second)),
-        },
-    }
-}
+pub use self::snapshot::{
+    capture, DirectionSnapshot, LayoutSnapshot, SessionSnapshot, TabSnapshot, WorkspaceSnapshot,
+};
+#[cfg(test)]
+use self::snapshot::{capture_node, PaneSnapshot};
+use self::snapshot::{parse_snapshot, snapshot_file_version, SNAPSHOT_VERSION};
 
 // --- Restore ---
 
@@ -601,17 +353,6 @@ pub fn clear() {
     info!(path = %path.display(), "session cleared");
 }
 
-fn parse_snapshot(content: &str) -> Result<SessionSnapshot, String> {
-    let raw = serde_json::from_str::<RawSessionSnapshot>(content).map_err(|e| e.to_string())?;
-    if raw.version > SNAPSHOT_VERSION {
-        return Err(format!(
-            "snapshot version {} is newer than supported {}",
-            raw.version, SNAPSHOT_VERSION
-        ));
-    }
-    migrate_snapshot(raw)
-}
-
 pub fn load() -> Option<SessionSnapshot> {
     let path = session_path();
     if !path.exists() {
@@ -627,10 +368,10 @@ pub fn load() -> Option<SessionSnapshot> {
     match parse_snapshot(&content) {
         Ok(snap) => Some(snap),
         Err(e) => {
-            if let Ok(raw) = serde_json::from_str::<RawSessionSnapshot>(&content) {
-                if raw.version > SNAPSHOT_VERSION {
+            if let Some(version) = snapshot_file_version(&content) {
+                if version > SNAPSHOT_VERSION {
                     warn!(
-                        file_version = raw.version,
+                        file_version = version,
                         supported = SNAPSHOT_VERSION,
                         "session file is from a newer herdr version, ignoring"
                     );
